data-process/requirements.txt
osgeo
folium
geojsonio
geopandas
geopy
Pillow
rasterio
sentinelsat
shapely

